BTD: goods used in advance inside the propagation

+++ toulbar2/tb2store.hpp
@@ -122,7 +122,7 @@ public:
             }
             
             index = y - 1;
-            base = (long) pointers[y];
+            base = (int) pointers[y];
         }
     }
 };
